When running actions, ensure that function action's function has finished before executing the next action. This provides a stronger assurance that actions run in the expected order.
Consider a case where the user had specified the following actions:
- function action A
- end_conversation action B with a goodbye message to be spoken before the bot leaves

Prior to this change, it was possible that B's goodbye message would be spoken *before* A's function runs, because the enqueued `TTSSpeakFrame` would arrive at the TTS frame processor before A's `FunctionActionFrame` reached the pipeline sink (and before the sink-watching task then picked it up and fired the "on_frame_reached_downstream" event).

This change also ensures that actions run in the specified order even though event handlers (like "on_frame_reached_downstream") are wrapped in asyncio Tasks, which run in a nondeterministic order.